requirements.txt
google-adk==1.25.0
litellm==1.66.3
pydantic==2.12.5
python-dotenv==1.1.0
yfinance==0.2.56
langfuse
openinference-instrumentation-google-adk
psutil
rich